separated the chart options, sorting logic, data processing from Request Limit component
